guiApp/CanBusSnnifer/requirements.txt
PyQt5==5.15.2
pyqtgraph==0.11.1
pyserial==3.5
numpy==1.19.5
cobs